requirements.txt
Django==4.2.0
psycopg2-binary==2.9.6
gunicorn==20.1.0
django-environ==0.10.0
djangorestframework==3.14.0
django-filter==23.1
django-cors-headers==3.14.0
pillow==9.5.0
whitenoise==6.4.0
